backend/requirements.txt
ultralytics
opencv-python
supabase
python-dotenv
fastapi
uvicorn
python-multipart
twilio
